世代ごとの処理をモジュールに切り出した
